Started work on C++ implementation
It was relatively easy to write quick-and-dirty code for this in JavaScript, but it proved somewhat harder, at least for me, to write clean code. My goal was to define each field as a "bitnumber" that knows its bitcount and other information and can feed that information to a bitstream object largely behind the scenes, i.e. I wanted to be able to specify the bitcount and other associated properties just once in a struct's constructor and then have the stream's read method (or a read method on the bitnumber) do the rest. This proved difficult to get right.

I've attempted the "clean" approach I had in mind in C++. It needs a lot of cleanup (prototyping was a messy process), but it works: it can read at least as far as the hidden Loadout Camera Time option under the (mpvr) block's Respawn Options. This is significantly further than my JavaScript attempt at a clean approach was able to get.

Some terminology, then:

bitnumber: A field that saves to a specific number of bits. A bitnumber may have a prefixed sentinel bit indicating the presence or absence of further data (e.g. the Jump Height option: if the sentinel bit is falsy, then the option isn't present and no more bits should be read).

bytenumber: A bitnumber whose bit count matches that of a typical variable (i.e. 8, 16, 32, 64), and that lacks a sentinel bit and any other special encoding features.

bitbool: A bool that serializes to a single bit.

bitstream: An object that reads one bit at a time.

bytestream: An object that reads one byte at a time. This exists as a different object from (bitstream) but with a semi-comparable interface, as some structs occur both in byte-aligned and bit-aligned data blocks (i.e. game variant header data which appears in both chdr and mpvr).