Map permissions, player rating params, TU1 options
Added code to read a game variant's map permissions, player rating params, option toggles, and TU1 options. (This includes a member function on cobb::bitstream to read the compressed floats that Reach uses for TU1 options.) We now have parity with the JavaScript prototype.

Brought in cobb::bitset to make it easier to load option toggles.