Megalo condition work
Option toggles: fixed length of bit chunks; it's supposed to be 32-bits.

Miscellanous fixes to the Megalo opcode argument system.

Started work on trying to parse Megalo trigger conditions from a file. Currently it's not going well; the first conditions are garbage data, but not so garbage that they trip my error-checking.

Brought in my string-helper functions.