Debug helpers
When compiled in debug, ReachGameVariantOptionToggles keeps track of what bits were set when it loaded, in an easy-to-view vector of indices.

Testing with a TU One Flag variant confirms that ReachGameVariantOptionToggles is loading properly. The only explanation i can think of for Megalo conditions being broken is that we're loading them in the wrong place, but the spot I picked matches what I'm using as a reference...