Condition progress
We can now stringify the values of MegaloValueEnum-listed enums; both SimpleValues and ComplexValues use this.

Fixed _getComplexSubtype; it no longer treats the type index as signed.

Apparently the bitlengths for some of the flags, and possibly some of the enums as well, are wrong. Splendid. Raising killer_type from 3 bits (exactly enough to hold all of its values) to 5 (used for that flags-mask by Horizon's WumboScript source code) was enough to allow conditions to parse a little further.