Conditions: index type stringifying
Index types can be stringified, though the code is obviously very rudimentary (i.e. just displaying the index type and the index value, instead of, like, actually going and getting the name of the thing being referred to e.g. the name of an option, the name of a set of player traits, etc.).