Action fixes
Fixed the "CHUD Message" opcode. Headhunter (vanilla and TU) and Race now load properly as a result.

Fixed format strings for the "Set Respawn Object For for Team" and "Set Respawn Object for Player" actions.

Turns out, the "team or player" enum doesn't allow "none" as a value; rather, that value means "all players." I was able to glean this from examination of Headhunter's script, where the "destination moved" broadcast went out to "none" players.

Further testing shows that as of this commit, Rally, Skeeball, and Slayer work.

Fixed the list of testing files.