Gametype resaving
We can now write out the full contents of the (_blf) and (chdr) blocks. I was having some struggles with endianness but I think I have them somewhat resolved.

The cobb::bitstream class reads everything in big-endian by default, since that's the format that bit-aligned values have to be in. The cobb::bitwriter class has now been altered to write byte-aligned values in little-endian by default and bit-aligned values in big-endian; when writing a byte-aligned value, you can specify big-endianness if needed (it's needed for chunk header data, and for short 0xC in the blam chunk).