MPVR saving work
I've started work on the code to save the MPVR block, but there are errors (defined here as inconsistencies with the original file as viewed in HxD) both fairly early in the saved block and throughout its later portions (apparently including the game variant's misc options).

It's worth noting, also, that the code to save a string table is extremely primitive at present. It's impossible to save a game variant without some implementation of overlapping strings within the string tablem, as the team name tables are so small that they don't even support translations (they basically all have to be the same string). With that implemented, however, we should strive to improve this further in the future.

Fixed interactions between cobb::bitwriter and cobb::bitnumber: the latter now properly handles its offset and its presence bit when saving. Fixed default endianness on bit-aligned wstrings as handled by cobb::bitwriter.