Progress on resaving
Fixed a mistake in cobb::bitwriter::_write that caused us to lose previously-written bits when writing values with certain alignments.

Simplified presence bit handling in cobb::bitnumber.

Fixed a mistake in the Megalo::Trigger::to_string code, and changed an action's format string in light of its value not being displayed as a bool.

Comparing by eye, we are now accurate up to 0x3FF.