Resaving mostly done
Fixed heap corruption which occurred when writing Megalo conditions. I'm not sure exactly why it occurred other than that it was a defect in how cobb::bitwriter accessed a byte to write to it.

With this change, we now serialize all script content, and all of the data that we serialize is accurate to the original file using a by-eye comparison. (We do not, however, attempt to match the massive amount of padding at the end of MPVR.) It's a huge amount of data, so we will want to do an automated comparison later.

Remaining tasks: add support for blocks appearing in variable order (needed for the _eof block); make it possible for blocks to update file lengths and block lengths in their data; and add explicit support for the _eof block (since it contains a file length in little-endian).