Trying to figure out SHA-1 rehashing
Pretty sure we're hashing the right data and handling endianness properly. What I'm not sure of is whether we're hashing the right length (a length gets embedded with the data to hash).

Fixed writing of the encoding and engine versions for MPVR. Added methods to cobb::bitwriter to enable us to write blam header chunk lengths and file lengths; we currently write the chunk lengths; we do not yet write the file lengths.