Deleted unused files; set up Qt
I had to recreate the main project from scratch, but it's now Qt-enabled. I don't see any options to control whether it links statically or dynamically (i.e. Qt DLLs) but I think the latter is the default, so... :\

Currently it displays a "dead" window with some non-interactable widgets and whatnot. Also, it only does that if I disable all of my testing code. I do not know why. So that's great. I'll have to focus on the UI for now, and once I have that reasonably functional I can have it call into the Reach-related code.

I think the best option for now would be to do a Creation Kit-style treeview on the left, and a pane with switchable pages (based on the treeview) on the right. It'd allow faster navigation than a breadcrumb menu and also Qt doesn't seem to have a breadcrumb menu widget built-in anyway. So I need to figure out how to make the "Open" option in the File menu open a file, and then populate the treeview based on what's in the file (e.g. scripted player traits, etc.).