Main window: fixed alignment; started treeview, options
The tree-view is now a tree-widget and can be used for navigating to any implemented pages. Added a page for General Settings. Fixed alignment/layout for the main window.

I should start work on making it possible to edit things via the UI. I almost wonder if it might be best to create templated versions of the various UI classes which take pointer-to-members or something. I need to remember NOT to allow any edit-related code to fire while (isUpdatingFromVariant) is true; apparently, when we make changes to UI controls through code, that sends "change" events same as the user making changes, so we need that bool to avoid recursion or wrongly flagging things as "edited" when we set the controls' initial state after loading a game variant.