Player traits, option toggles, Save As fix
Implemented UI for player traits, though I've not yet tested to ensure that edits are saved properly.

Implemented UI for option toggles. Currently testing.

Fixed "Save As," which broke in a regression due to a copy-and-paste mistake in the game variant content header code. Fixed an unrelated bug where running "Save" after "Save As" would write to the original file rather than the newer one. Made it so that by default, the "Save As" dialog opens with the name of the original file.

Added value 5 to Player Traits: Active Camo, as it is used by the active camo powerup traits.