Saving is finished; all authorship data is alterable
Our saving code now writes file lengths to the places they need to be written within the file. As part of this, we now handle the (_eof) block directly.

The UI now allows you to set the "created" and "modified" dates on the file, and to clear the authorship XUIDs.