Moved MP variant data from GameVariant to data class
All multiplayer variant data is now stored in the GameVariantDataMultiplayer class.

The editor now fails with an error message when attempting to open a Firefight variant. The messaging could be improved -- it doesn't explain that we don't support Firefight variants, for example. I should probably define a thread-local variable somewhere to hold loading state, including encountered errors, so that callers can identify the specific details of a failure to load.