Improved load error handling
All load errors, including in Megalo, are handled by a new thread-local singleton, GameEngineVariantLoadError. (Is it still a singleton if it's thread-local? That's one instance per thread.) This singleton is used to provide a detailed error message via message box.

The cobb::bitreader class no longer asserts when it tries to read from EOF.

TODO: Classes need to double-check whether the stream is still in-bounds when reading, and fail if it isn't. (Better yet: we should have a way to check by how far we've overshot EOF, i.e. how many bytes/bits we failed to read; if that's non-zero then fail. This may be a better approach because if we're AT EOF after all the data, then we still read everything we needed to.)