Main window polishing is pretty much done
Moved loadout palettes out of the main window and into their own file.

Overhauled the main window's navigation code; navigation is now generated at run-time. For now this allows us to only show Forge options in the navigation when viewing a Forge variant; if Firefight variants get decoded, we'll be able to use this to build a totally different nav tree for Firefight. This change also knocks out a few other to-do list items, like preventing the user from viewing inoperable player trait and team config forms when no file is open, and it means that we're not doing tons of string comparisons anymore to find out what the user clicked (there was no way to add user-data to preplaced QTreeWidget items in Qt Designer; the only identifying information for any preplaced item was the label).

Added a "Welcome" page, which is displayed when no file is open.