Gametype engine category fixes
Fixed the loading of signed bitnumbers that use offset values (i.e. they're signed but the only allowed signed value is -1, and they're incremented during serialization such that the actual encoded number is unsigned). This previously broke the "Action Sack" engine category index (16 in  a 5-bit number, which we incorrectly loaded as -16).

If you load a game variant with an unofficial engine category index, we correct it to Action Sack (so that it shows up in MCC's menus) and alert you.