Experiments with script syntax
I have a copy of Alpha Zombies' script exported to a plain-text, plain-English syntax. On a whim I decided to translate that content to something scripty, making up a syntax as I went, and I think I quite prefer the result.

A scripting language would be faster to work with; you'd be able to just type stuff up, though at the risk of running into syntax errors. However, I'd have to put in the work to properly parse it (the current code creates a hierarchy of code blocks correctly, but doesn't actually parse individual statements yet), and to add sensible error handling to the parser.

A UI-based approach would be much more on-rails, but it would require that I write code to transform the in-memory data (with a script-based approach, I'd just discard the in-memory data and generate new data from scratch based on the entered script code). I dunno. We'll see.