Fixes to string table export
We now collapse duplicate strings across the string table when saving, which is a step beyond what Bungie does and which should be consistent with KSoft.Tool. This change should allow game variants with especially large script blocks to serialize correctly; in tests, "SvE Mythic Slayer," which doesn't work in v1.0.1, is able to properly fit within the file size limit. Once I've confirmed that I'm not doing anything else wrong in regards to that file, I'll test and eventually enable more optimizations to string tables.