Score to Win; cobb::reference_tracked_object; Forge label improvements
Added "Score to Win" to the UI after having forgotten it for the initial release.

Created cobb::reference_tracked_object, a generic superclass usable for keeping track of what objects are referring to what. Modified Forge labels to use this system (currently, opcode arguments work with it; trigger for-each-object-with-label values do not).

The motivation here is to rely less on indices. Previously, if I wanted to remove or reorder the Forge labels (say, in a script editor), I would have to iterate over every single trigger and every single trigger opcode and perform fix-up operations on all Forge label indices. Once this system is fully in place (again, triggers themselves need to be made to work with it), I won't need to fix up indices all over the data; instead, I just need to make sure that each Forge label knows its final index on save (to accommodate the save code for trigger opcodes, which can't "see" the broader game variant and therefore can't check a label's index in the label list on its own).