Started work on editing scripted options
Created the form to edit scripted options. I'm pretty sure we get string picking/editing "for free" since so much of that is handled by the string picker itself, but we still need to double-check to make sure that the option UI responds properly to an option or one of its values being renamed. We also need to set up editing of all other fields; currently the UI is read-only beyond what we get for free.

Game variants now store triggers in a cobb::pointer_list<Trigger> instead of a vector of unique pointers; this means that we can pass [references to] the trigger list around (e.g. to Trigger::to_string) without the recipient having to care what kind of pointer is in the list.