v1.0.3
Official release for v1.0.3. Note that 90% of the work that's actually in this commit isn't in the 1.0.3 binary, because the script editor is only even accessed by anything when the program is compiled in Debug.

Player traits can now be added, reordered, and removed, though I still need to test this (MCC is bugged -- it only shows up to four sets of traits, and this affects the official Halo Chess so I know it's not my fault -- so my original test did not go as planned). The main window now properly maintains the current navigation when traits are renamed, added, or deleted (though reordering the currently-focused trait set can confuse the main window).

The "player traits" opcode argument now refers to the targeted traits via a cobb::reference_tracked_object::ref.