Script metadata strings; team names
Script metadata strings are now properly editable (though we can't add strings that are missing, which is a minor problem).

Team names can now be edited (though we can't add custom team names when they aren't already present, which is a minor problem). To facilitate this (and the prior change), the localized string editor dialog can now be told that the string it's editing isn't part of the standard string table (i.e. so it doesn't fire gametype-wide change events or offer a "Save as New" option).

Player rating parameters can now be edited.