Lua-style parser
We can now properly parse every for-loop type except for "for each object with label X" loops.

We can now parse "or" and "and" keywords.

Made some preparations for if-statements. We're going to store the if-condition (delimited with parentheses) as an MBlock instance with no nesting support, as a member on the MBlock instance that represents the if-block itself.