Lua-style parser: string literal and for-loop support
Forked the parser into "parser2.js" and added support for string literals. We now fully process for-each-object-with-label loops as well. The new parser model uses a state object, so that logic for token extraction and keyword handling can be separated out of the main parser loop while still being able to influence parse state.