Lua-inspired parser: minor cleanup
Took a piece of code that was duplicated and split it off into a function that gets called from two places. Added a to-do list to the top of the file.

At this point, we have everything but alias and expect declarations. Once those are done, we can test on the full Alpha Zombies script.