Script parsing prototype: first-pass on Alpha Zombies
Alpha Zombies now succeeds at first-pass parsing (with some fixes to syntax errors in the script itself). Bugfixes include:

Within function calls with multiple arguments, commas are effectively expression joiners and are now parsed as operators to effect this.

Fixed MExpression.insert_item: the check for "func(a).b" was incorrect and was accidentally testing for the "." in (last) rather than the first in (item).

Attempted to fix bugs in line number tracking. However, I think perfect tracking is a lost cause; we'd be better off tracking the *last-seen* line number and its position from (tryExtractExpression) and then, any time we need to know the *current* line number, just counting all line breaks from that position. The basic problem here is that we have multiple loops running on the input stream which can all advance the stream as needed according to their own logic.