Lua-inspired parser: fixes, serialization
Fixed an unclosed function call in the Alpha Zombies test script.

Added code to take the content found by the first parser pass and write it back out as a string; I now need to take a look at that output and ensure that it matches the original script (a blind automated comparison will fail on comments, whitespace, and other superficial differences). Once I'm sure that it does, we need to move on to writing the second parser pass.