Lua-inspired parser (alternate)
Started drafting an alternate approach to parsing.

The current first-pass parser is capable of reading statements that we can't actually compile. For example, if we want to compile {a += b * 3}, we need to be able to select a variable to hold the temporary (b * 3), and I honestly think that that's something that should be left to the human script author instead (since we have only a limited number of variables available, more akin to registers). We could still compile any expression that used consistent operators or more generally operators that can be executed in order (e.g. {a += b - 3 + 4}), but I don't want to create a situation where it isn't 100% clear what can and can't be compiled, y'know?

So I've started work on an alternate, simpler first-pass parser. Compiled Megalo code only allows four kinds of statements: blocks; comparisons; assignments; and calls. Comparisons can only appear within if-conditions; assignments can only appear outside of them; and calls can appear as statements in either setting, and as the righthand side of an assignment (but not a comparison). As such, parsing this should be rather simple (though my current parser is an absolute nightmare):

If we're not in a statement, then the next token must be a "word," i.e. a variable or keyword. If it's a keyword, then we handle it accordingly; otherwise, the token after it must be either an operator (denoting a comparison or assignment) or an opening parentheses (denoting a function call). Once we've found either, we know what kind of statement we're in and can parse further.

Our current draft of the simple parser... doesn't function this way. It should be altered or rewritten.