"Simple" Megalo parser: cleanup, line comments, docs
Added documentation comments to further clarify the flow of execution/parsing. Removed some unused variables from the loop.

The parser should now understand "--" as the marker for a line comment.

If the lefthand word in a statement begins with "-" and is not a number, that will now be flagged as a syntax error.