Simple Megalo parser: for-loops, variables
Implemented MVariableReference.extract. Currently this is only used for MFunctionCall.context and not for variables referred to by MAssignment or MComparison.

Removed the unused "extract" member functions from MSimpleParser.

MBlock.insert_item now properly sets up parent/child relationships and returns the inserted item.

Split the logic for each statement type (and for statement-start handling) out of MSimpleParser.parse and into separate member functions, so that the higher-level logic is somewhat easier to follow.

Added MSimpleParser.scan as a function to loop over characters, keeping track of comments and whitespace, and modified MSimpleParser.parse to use this function. The benefit of this is that if we do want to extract tokens of a specific type (e.g. particular words), we can now do so without breaking our tracking of comments or line numbers.

We now handle all for-loop types except for for-each-object-with-label loops.