Simple Megalo parser: conditions
MSimpleParser.parse should now only be used for blocks of code. Parsing of actions/blocks has been split off from parsing of conditions, though some functionality overlaps.

The "simple" Megalo parser can now read if-blocks and the conditions therein, including "and" and "or" joiners. It's a bit messy -- some state variables that are shared between reading conditions and reading actions/blocks -- but it works.

The following errors are caught:

if a then -- syntax error: "word word"

if a == b or or c then -- syntax error: "word word"

if or a == b then -- syntax error: joiners must come after a condition

if a == b do -- syntax error: you must use "then," not "do"