Simple Megalo parser: for-each-object-with-label; EOF fixes
Finished implementing for-each-object-with-label loops.

We no longer fail to parse tokens that end at EOF. MSimpleParser.scan pretends that all text streams end in a single whitespace.

We now detect whether there are any unclosed blocks or incomplete statements at the end of the text stream, and fail with appropriate error messages.

If we detect an out-of-place elseif- or else-block, we now detect the possibility that a script author incorrectly placed an "end" token before these (i.e. "if x() then y() end else").