Simple Megalo parser: token positions; to-do list
We now attempt to store the starts (as stream offsets and line/col numbers) and ends of each parsed object.

Added some more tasks to the to-do list: we need to disallow keywords in variable names, and we need to use "real" parsing for function arguments so that we don't mishandle commas inside of string literals.