Simple Megalo parser: simplified the parse logic
With the way _parseFunctionCall is designed, it doesn't need to be run within the main parsing loops anymore; it's designed to loop and extract tokens on its own, kind of like what the handler for the "for" keyword does, so it can just be called the instant we encounter a function call's opening parentheses.

Rearranged some function in the source code so that related functions are next to each other.

Forgot to note this in the last commit, but I fixed extractIntegerLiteral so that it no longer advances to one character past the end of the integer.