Simple Megalo parser: fixed line number tracking
Modified MSimpleParser.scan so that if it sees the same line break twice in a row, it doesn't increment the line number counter twice.

Modified all callers of MSimpleParser._parseFunctionCall to avoid accidentally skipping the glyph (often a line break) immediately after a function call. Added documentation comments to that function to explain what's needed to avoid the problem.

Line numbers are now properly tracked for the entire Alpha Zombies test script.