Megao parser: Rewriting variable resolution code
I'm rewriting the variable resolution code to use an OO approach. Current code is messy and transitional.

The new approach is to define objects that represent script typenames, script namespaces, members within those namespaces, and properties on those typenames. We then resolve things one MVariablePart at a time, storing the analyzed data on each part. If resolution completes successfully, then we can examine the parts to figure out the final analyzed values on the MVariableReference (this step is not implemented yet).