Megalo parsing: alias resolution
Resolution of alias declarations is basically done now.

Alias invocations now properly trap (namespace.var.var.aliased_var). The problem was that when we spliced the contents of an alias in, we spliced them raw -- the original objects, with ownership pointers pointing elsewhere. We needed to duplicate the contents of the alias and properly update ownership on the duplicates, in order for those contents to then resolve properly as variables.

Removed a ton of now-unused code -- previous attempts at implementing variable/alias resolution.