Megalo parser: fixed some MVariableReference ownership
We now track MVariableReference ownership for assignment statements (the "source" half) and for integer function call arguments.

If there's anywhere else we've missed, we're guaranteed to catch it in C++; I plan on making every (owner) field on these objects a reference, which means that it will HAVE TO BE initialized as a constructor argument.