Megalo parser: scope/which/index work
Started work on having MScriptVariables identify their types, scopes, whiches, and indices. Needed to move several enums across files to account for the order in which files are loaded.

Object and player variables have the code to identify their (scope), (which), and (index) values. We still need code for numbers, teams, and timers.