Megalo parser: number/timer/team scope/which/index
Code to resolve scope/which/index for numbers, teams, and timers is now complete. To make this code easier (i.e. to avoid having to switch-case a ton within the resolve functions), we've added fields to MScriptProperty and MScriptNamespaceMember indicating: the proper (scope) values to use when the property/member is a number or timer and is being accessed; or the proper (which) values to use when the property/member is a nestable variable and it or a nested variable/property is being accessed.