Megalo parser: event handler syntax
The first-pass parser can now parse event handler declarations, e.g. (on event name: do ... end). The ":" glyph is now classified as a "syntax char," which is necessary to prevent extractWord from "eating" it when reading the event handler name.

Did a bit of cleanup: MSimpleParser.getHandlerForKeyword didn't need to exist; it was only called from MSimpleParser._parseActionStart so I've inlined it there.