Started work on the script decompiler
The C++ program now has a work-in-progress decompiler in the script editor window, under "Gametype Code." Currently the decompiler only writes the script's block structure (including proper handling of if-blocks and conditions in the middle of triggers); all opcodes are written out as the word "OPCODE". In order to decompile opcodes, we will need to add support to every opcode argument type, and we will need to add opcode-function-to-script mapping data (also needed for parsing).

Fixed promoted widget paths throughout the project. When editing promoted widget classes in Qt Designer, the designer likes to show promoted widgets for either all open UI files or all UI files in the project (not sure which), but in reality promoted widget definitions exist on a per-UI-file basis, and the header paths need to be relative to the UI file's path.

Made similar fixes to some include paths in my own code files. (I don't know why the files with bad include/promotion paths all compiled before; I can only assume that there are some serious bugs in Qt's build tools, more likely in MSBuild, or in the integration between the two.)

Fixed the loader code for object-variable opcode argument values; they now use MegaloVariableScopeObject.which_bits() instead of the player which-bits (luckily, it seems they both have the same number of which-bits, so run-time behavior was still correct).