Started fleshing out OpcodeArgTypeinfo
OpcodeArgTypeinfo is fleshed out enough for me to start writing instances. However, I've run into a bit of a snag: some multi-part argument types are a bit untenable; as an example, the "shape" opcode type consists of an enum followed by a variable number of number values. If we keep these arguments as whole arugments, then we lose out on some of the things we wanted to do, like consolidating the handling of all enums, AND the handling for multi-part arguments will have to be able to parse variable names; conversely, if we split these argument types apart, then we have to be able to treat opcodes like "set object shape" as varargs -- the code to compile a function call needs to, when compiling arguments in the call, "consume" one or more arguments at a time and return the number consumed.

Took all this as an opportunity to move several OpcodeArgValue subclasses' methods out of their header files. We're defining the typeinfo instances outside of the headers as well so that if we extend that class (as we're very likely to do in the future, and as we WILL do for variable types), we don't have to recompile like 90% of our Megalo stuff again.

Fixed the decompiling of opcodes with results; we were accidentally writing "result. = whatever" instead of "result = whatever".