Started work on redesigning how we keep opcode arguments in memory
For now, I'm just recreating all of the opcode stuff in a separate folder and namespace; this will include conditions, actions, and the condition- and action-bases. When everything's converted over and it all works, I'll just swap it all in.

This includes a *very* preliminary implementation of cobb::bitvector64 which is currently incorrect, and a preliminary implementation of DetailedEnum (which I'll eventually need to move out of the gametype-specific folders).