Further planning for the new opcode arg system
We need a 128-bit buffer for every opcode arg value, as the largest possible argument (a box shape with integer constants for all four dimensions) is 90 bits.

We need to redesign the opcode arg typeinfo functors so that they work directly with the 128-bit buffer and the array of reference-tracked-object refs; this is needed so that functors can call each other in the case of "nested" types (e.g. the number-variable data inside of shape arguments).