Opcode arg revamp: more plans
I realized that we're missing another piece of the puzzle: index fixup.

The reason we use cobb::reference_tracked_object in the first place is to deal with, for want of a better term or really any consistent terminology at all, "indexed references." Several opcodes refer to list items by index, such as script options, player traits, and so on. This creates two needs: we want to be able to reorder these lists without breaking all of the opcodes which refer to their contents; and we need to know when a list item is in use (whether by an opcode or something else), so we don't allow the user to delete it.

The naive solution to this problem is to just manually fix up every opcode every time we reorder anything in any list, but that's... horrible. It's horrible and it's error-prone. The solution we currently employ is to create two interfaces: cobb::reference_tracked_object and cobb::reference_tracked_object::ref. The former is a superclass/interface to be applied to any object for which we need to track inbound references (e.g. a script option referred to by an opcode) or outbound references (e.g. an opcode which refers to a script option). The latter is what a R.T.O. must use to actually refer to another R.T.O.: it is designed to be aware of its containing struct [1] and to use smart-pointer behaviors to manage inbound and outbound references.

This solution, of course, introduces some complications for opcode arguments. My original plan was to just load the bits for an opcode argument and then, when resaving, blindly write them into the target file; this will obviously not work if the opcode refers to a reference_tracked_object that has been reordered within its containing list. The old opcode argument system has fixup code and we're going to need some sort of fixup system here as well, to be performed whenever we resave the opcode.

My current preference is to make two changes: ditch the reference_tracked_object interface in favor of a handle/refcount system similar to that used for in-world objects in Skyrim, where handles are stored globally so that objects with handles don't have to know their owning game variant; and then, now that we don't have to deal with the limitations of smart pointers that are aware of their owners, make (arg_rel_obj_list_t) an array of structs, each consisting of a handle, bit offset of the index, and bitcount of the index (all to be set when loading or compiling). This will allow for a generic index fixup approach.

[1] Now that I think about this, I don't think that the way this works is strictly compatible with having arrays of the things, so we may need to revise all this anyway.