Change to reference-tracking
We no longer use cobb::reference_tracked_object to know whether certain objects are referred to by the gametype script (or other parts of the gametype in general), as this was total overkill and it imposed challenging constraints on the opcode argument refactor. We instead use a simpler refcounting system.

I ran two tests on this change. First, I resaved Invasion: Boneyard and confirmed that the result was byte-for-byte identical to a resave from the program's release build. Second, I saved a copy of Invasion: Boneyard with the script options, script traits, Forge labels, HUD widgets, and other data reordered at random, and confirmed in-game that the gametype still behaved properly.