Opcode arg redesign: index fix-up progress
The (arg_rel_obj_list_t) type is no longer an array of pointers, but rather a struct containing an array of pointers and one bitrange for each array. This is in service to the index fix-up design.

The two currently-defined types that use indexed-refcountable objects, forge_label and player_traits, should now properly provide bitrange data. All that's left is to write code to actually perform index fix-up when needed; we can do this with a member function on OpcodeArgValue.