Opcode arg redesign: fragment-specifiers
Expanded the typeinfo functors' (fragment) argument into a small struct called the "fragment specifier," which can contain a bit-offset intended to help nested types invoke decode functors without having to copy and shift the opcode argument's binary data (this would be necessary if it weren't possible to tell a functor to start from an arbitrary offset within the binary data).

All currently-implemented types in the new system have been modified to use the fragment-specifier's bit offsets and indices in all applicable functors.

Added a helper function to pull an index out of a bitarray given the bitrange of that index; this is primarily useful for postprocess functors but it can also be used by decode functors when a relevant-object is missing (which, granted, should never happen).