Opcode arg work: started on variables
Started work on mapping script variables into the new system.

I have what should be a fully-functional setup for timer variables, but as you can see from looking at it, it's very repetitive -- not necessarily the same things being done in several places, but very closely related things. I want to take the type/scope/which/index system and make it generic -- make it possible to define the different "scope" values for each type as data, and have helper function which then act on that, so that we don't have so much quasi-repetition. It's basically going to be a "must" when we get to doing number variables.