Opcode arg work: variable work
Still working on a generic approach to variables. Once the approach itself is set up, I need to actually create variable types that use it.

Noted down a design oversight in the IGameVariantDataObjectNeedingPostprocess idea. We'll tackle it after all opcode argument types are defined in the new system; we just need to make it so that load functors have access to the containing game variant.

Beginning to feel increasingly unsure about the OpcodeArgValue system itself -- the decision to just have a single OpcodeArgValue class and do most of the work in functors. I think it'll be easier for me to compare the two approaches once I have compiling implemented, and I do think that it'll at least be easier to implement (if not maintain) compiling in the newer approach. If I ever need to switch back to having a class and a typeinfo-instance per opcode argument type, it'll be easiest to do so once I have everything implemented in the new system: just roll back the files and backport stuff, right?