Opcode arg work: variable work
Wrote the code for generically loading a variable, and the code for generically postprocessing a variable that refers to indexed data.

VariableScopeIndicatorValue no longer uses a functor for the index bitcount when the indicator is for indexed data; a functor shouldn't be necessary because indexed lists always use the same bitcounts (i.e. if the indexed data in question is, say, script options, then the bitcount will always be the same). We just store a bitcount member on VariableScopeIndicatorValue.

VariableScopeIndicatorValue::postprocess_functor_t now just returns the indexed object, given a variant and an index.