Opcode arg work: variable work
Fixed how global variables are handled when the variable type is enum-based (i.e. objects, players, teams); in that case, the "which" indicates the variable and the "index" is unused. Previously it was the other way around, which would not have allowed us to stringify/decompile these variables in a generic system.

Added GameEngineVariantLoadError-based error reporting to the generic variable load code. Error reporting is incorrect for "biped" property access and (as before) nonexistent for "team" property access.

VariableScopeIndicatorValueList instances are now defined for all variable types except "timer." OpcodeArgTypeinfo instances are now defined for all variable types.